2 25 746

+++ PA2/Graph.h
@@ -23,7 +23,7 @@ private:
     // (adjacent, weight)
     vector<vector<pair<int, float>>> adj;
     vector<int> parent;
-    vector<int> d;
+    vector<float> d;
     priority_queue<pair<int, int>, vector<pair<int, int>>, greater<pair<int, int>>> Q;
     // paths
     vector<vector<int>> paths;
